feat(proving): feature-gated prover adapters with generic Groth16 core
Split groth16.rs into generic core (Pairing-generic) and groth16_bn254
(BN254-specific snarkjs JSON serialization). Add Cargo feature flags:
- groth16-bn254 (default): BN254 Groth16 with snarkjs-compatible output
- groth16-bls12-381: BLS12-381 Groth16 via ark-bls12-381
- plonk-bn254 (default): halo2 KZG PlonK on BN254
- groth16-core: shared arkworks deps activated by any groth16 variant

All modules feature-gated in lib.rs. No-default-features compiles clean.
