+++ proving/Cargo.toml
@@ -10,6 +10,31 @@ authors.workspace = true
 keywords.workspace = true
 categories.workspace = true
 
+[features]
+default = ["groth16-bn254", "plonk-bn254"]
+
+# Groth16 core (generic over Pairing) — activated by any groth16 variant
+groth16-core = [
+    "dep:ark-groth16",
+    "dep:ark-relations",
+    "dep:ark-std",
+    "dep:ark-serialize",
+    "dep:ark-ff",
+    "dep:ark-ec",
+    "dep:ark-snark",
+    "dep:sha2",
+    "dep:rand",
+]
+
+# Groth16 on BN254 (Ethereum-compatible, default)
+groth16-bn254 = ["groth16-core", "dep:ark-bn254"]
+
+# Groth16 on BLS12-381 (Zcash/Filecoin-compatible)
+groth16-bls12-381 = ["groth16-core", "dep:ark-bls12-381"]
+
+# PlonK/KZG on BN254 via halo2
+plonk-bn254 = ["dep:halo2_proofs", "dep:rand"]
+
 [dependencies]
 # Achronyme crates
 memory = { path = "../memory" }
@@ -20,17 +45,20 @@ vm = { path = "../vm" }
 # Serialization
 serde_json = "1"
 
-# Groth16 (ark-works ecosystem)
-ark-groth16 = "0.5"
-ark-bn254 = "0.5"
-ark-relations = "0.5"
-ark-std = "0.5"
-ark-serialize = "0.5"
-ark-ff = "0.5"
-ark-ec = "0.5"
-ark-snark = "0.5"
-sha2 = "0.10"
+# Groth16 shared (activated by groth16-core feature)
+ark-groth16 = { version = "0.5", optional = true }
+ark-relations = { version = "0.5", optional = true }
+ark-std = { version = "0.5", optional = true }
+ark-serialize = { version = "0.5", optional = true }
+ark-ff = { version = "0.5", optional = true }
+ark-ec = { version = "0.5", optional = true }
+ark-snark = { version = "0.5", optional = true }
+sha2 = { version = "0.10", optional = true }
 
-# PlonK (halo2/KZG on BN254)
-halo2_proofs = { git = "https://github.com/privacy-scaling-explorations/halo2", tag = "v2023_04_20", features = ["circuit-params"] }
-rand = "0.8"
+# Curve-specific (activated by groth16-bn254 / groth16-bls12-381)
+ark-bn254 = { version = "0.5", optional = true }
+ark-bls12-381 = { version = "0.5", optional = true }
+
+# PlonK/halo2 (activated by plonk-bn254)
+halo2_proofs = { git = "https://github.com/privacy-scaling-explorations/halo2", tag = "v2023_04_20", features = ["circuit-params"], optional = true }
+rand = { version = "0.8", optional = true }
