+++ proving/src/lib.rs
@@ -1,3 +1,11 @@
+#[cfg(feature = "groth16-core")]
 pub mod groth16;
+
+#[cfg(feature = "groth16-bn254")]
+pub mod groth16_bn254;
+
+#[cfg(feature = "plonk-bn254")]
 pub mod halo2_proof;
+
+#[cfg(feature = "groth16-bn254")]
 pub mod solidity;
